\name{estimateRE}
\alias{estimateRE}
\title{estimateRE}
\usage{
  estimateRE(modelSpecs)
}
\arguments{
  \item{modelSpecs}{list with all necessary components for
  estimation}
}
\description{
  This function estimates the BLUP for the RE-Part, given
  the results of the estimation of all parameters -
  \code{\link{optimizeParameters}}
}

