% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{funzione}
\alias{funzione}
\title{Something about the function ....}
\usage{
funzione(x = NULL, ...)
}
\arguments{
\item{x}{first argument}

\item{...}{further arguments}
}
\description{
Something about the function ....
}
\examples{

ll <- funzione()





}

