% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackageTemplate.R
\name{r_hello_world}
\alias{r_hello_world}
\title{A "hello world" function used to test the 
\code{\link{rcpparma_hello_world}} implementation.}
\usage{
r_hello_world()
}
\value{
A 3x3 matrix with 7 on the diagonal and zero elsewhere.
}
\description{
A "hello world" function used to test the 
\code{\link{rcpparma_hello_world}} implementation.
}

