% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_bothproducts}
\alias{rcpparma_bothproducts}
\title{Calls both the inner and outer product functions, returning a list.}
\usage{
rcpparma_bothproducts(x)
}
\arguments{
\item{x}{A numeric vector of length \eqn{n}.}
}
\value{
A named list with two elements, the scalar inner and the matrix outer.
}
\description{
Calls both the inner and outer product functions, returning a list.
}

