% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_hello_world}
\alias{rcpparma_hello_world}
\title{A "hello world" function using the Armadillo library.}
\usage{
rcpparma_hello_world()
}
\value{
A 3x3 matrix with 7 on the diagonal and zero elsewhere.
}
\description{
A "hello world" function using the Armadillo library.
}

