% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_innerproduct}
\alias{rcpparma_innerproduct}
\title{Calculates the inner product of a vector using the Armadillo library.}
\usage{
rcpparma_innerproduct(x)
}
\arguments{
\item{x}{A numeric vector of length \code{n}.}
}
\value{
The scalar \eqn{x'x}.
}
\description{
Calculates the inner product of a vector using the Armadillo library.
}

