% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_outerproduct}
\alias{rcpparma_outerproduct}
\title{Calculates the outer product of a vector using the Armadillo library.}
\usage{
rcpparma_outerproduct(x)
}
\arguments{
\item{x}{A numeric vector x of length \eqn{n}.}
}
\value{
The \eqn{nxn} matrix \eqn{xx'}.
}
\description{
Calculates the outer product of a vector using the Armadillo library.
}

