\name{r.package.template-package}
\alias{r.package.template-package}
\alias{r.package.template}
\docType{package}
\title{
\packageTitle{r.package.template}
}
\description{
\packageDescription{r.package.template}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{r.package.template}
\packageIndices{r.package.template}

}

\author{
\packageAuthor{r.package.template}
Maintainer: \packageMaintainer{r.package.template}
}

%\references{
%}


\keyword{ package }
\seealso{
\url{http://www.r-project.org}
}

\examples{
rcpparma_hello_world()
}
