% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackageTemplate.R
\docType{package}
\name{RPackageTemplate}
\alias{RPackageTemplate}
\alias{RPackageTemplate-package}
\title{An R package template illustrating the use of various packages including:
\code{\link[Rcpp]{Rcpp}}, \code{\link[RcppAramdillo]{RcppAramdillo}}, 
\code{\link[devtools]{devtools}}, \code{\link[roxygen2]{roxygen2}}, 
and \code{\link[testthat]{testthat}}.}
\description{
An R package template illustrating the use of various packages including:
\code{\link[Rcpp]{Rcpp}}, \code{\link[RcppAramdillo]{RcppAramdillo}}, 
\code{\link[devtools]{devtools}}, \code{\link[roxygen2]{roxygen2}}, 
and \code{\link[testthat]{testthat}}.
}

