% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/templateTest.R
\name{templateTest}
\alias{templateTest}
\title{templateTest-Function}
\usage{
templateTest(x = "hi")
}
\arguments{
\item{x}{character}
}
\value{
returns 'hi'
}
\description{
describe me
}
\details{
more details
}

