\name{RPackageTemplate-package}
\alias{RPackageTemplate-package}
\alias{RPackageTemplate}
\docType{package}
\title{
\packageTitle{RPackageTemplate}
}
\description{
\packageDescription{RPackageTemplate}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RPackageTemplate}
\packageIndices{RPackageTemplate}

}

\author{
\packageAuthor{RPackageTemplate}
Maintainer: \packageMaintainer{RPackageTemplate}
}

%\references{
%}


\keyword{ package }
\seealso{
\url{http://www.r-project.org}
}

\examples{
rcpparma_hello_world()
}
