\name{rpackagetemplate-package}
\alias{rpackagetemplate-package}
\alias{rpackagetemplate}
\docType{package}
\title{
\packageTitle{rpackagetemplate}
}
\description{
\packageDescription{rpackagetemplate}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rpackagetemplate}
\packageIndices{rpackagetemplate}

}

\author{
\packageAuthor{rpackagetemplate}
Maintainer: \packageMaintainer{rpackagetemplate}
}

%\references{
%}


\keyword{ package }
\seealso{
\url{http://www.r-project.org}
}

\examples{
rcpparma_hello_world()
}
